Create timelog_artist_mailer
Get list of artist from artist table
For each artist check in all timelogs if there is any "0" or no timelog filled by artist for yesterday or day before and send mail to him/her. 

If it is for day before yesterday then cc mail to production head and Head.